Tests/LibWeb/Text/expected/css/subgrid-bounds-auto-row-placement.txt

first: 0 0 30 10
second: 0 0 30 10
